+++ tests/test_suite.py
@@ -14,13 +14,8 @@ from time import sleep
 from unittest.mock import Mock, patch, MagicMock
 import logging
 
-# Setup path for imports
-plugin_dir:Path = Path(__file__).parent
-sys.path.insert(0, str(plugin_dir))
-
 # Config is already mocked by conftest.py
 from harness import TestHarness
-from load import journal_entry
 
 logging.basicConfig(level=logging.DEBUG, format='%(asctime)s - %(levelname)s - %(message)s')
 
@@ -28,10 +23,21 @@ logging.basicConfig(level=logging.DEBUG, format='%(asctime)s - %(levelname)s - %
 def harness() -> Generator:
     """Provide a fresh test harness for each test."""
     test_harness = TestHarness()
-    test_harness.register_journal_handler(journal_entry)
+    test_harness.set_edmc_config()
+
+    # Now we can import Router modules
+    from load import plugin_start3, plugin_app
+    import bgstally.globals
+    test_harness.bgstally = bgstally.globals.this
+    
+    plugin_start3(str(test_harness.plugin_dir))
+    plugin_app(test_harness.parent)
+
+    test_harness.register_journal_handler(test_harness.bgstally.journal_entry)
     test_harness.commander = 'Testy'
     test_harness.is_beta = False
-    
+    test_harness.system = 'Sol'
+        
     yield test_harness
     # Cleanup if needed
 
@@ -51,8 +57,9 @@ class TestFleetCarrier:
         events:list = harness.events.get('carrier_events', [])
         harness.bgstally.fleet_carrier.overview['carrier_id'] = 3709409280
         assert harness.bgstally.fleet_carrier.overview.get('carrier_id') == 3709409280
+        assert harness.config.get_str('BGST_Status', default='On') == 'On'
         harness.fire_event(events[0])
-        assert harness.bgstally.fleet_carrier.overview.get('jumpDestination') == ''
-        harness.fire_event(events[1])
-        assert harness.overlay.msgs != {}
+        assert harness.bgstally.fleet_carrier.overview.get('jumpDestination') == 'Bleae Thua ZE-I b23-1'
+        #harness.fire_event(events[1])
+        #assert harness.overlay.msgs != {}
